+++ examples/with-next-siwe-iron-session/CHANGELOG.md
@@ -1,5 +1,26 @@
 # with-next-siwe-iron-session
 
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies [ab051b9]
+  - @rainbow-me/rainbowkit@0.12.12
+
+## 0.0.30
+
+### Patch Changes
+
+- Updated dependencies [0469e00]
+  - @rainbow-me/rainbowkit@0.12.11
+
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies [ecaa85f]
+  - @rainbow-me/rainbowkit@0.12.10
+
 ## 0.0.28
 
 ### Patch Changes