+++ examples/with-next-siwe-next-auth/CHANGELOG.md
@@ -1,5 +1,29 @@
 # with-next-siwe-next-auth
 
+## 0.0.31
+
+### Patch Changes
+
+- Updated dependencies [ab051b9]
+  - @rainbow-me/rainbowkit@0.12.12
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.9
+
+## 0.0.30
+
+### Patch Changes
+
+- Updated dependencies [0469e00]
+  - @rainbow-me/rainbowkit@0.12.11
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.9
+
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies [ecaa85f]
+  - @rainbow-me/rainbowkit@0.12.10
+  - @rainbow-me/rainbowkit-siwe-next-auth@0.1.9
+
 ## 0.0.28
 
 ### Patch Changes